Converts text file to hdf5 data file

Converts text file to hdf5 data file

Usage

bdImport_text_to_hdf5(
  filename,
  outputfile,
  outGroup,
  outDataset,
  sep = NULL,
  header = FALSE,
  rownames = FALSE,
  overwrite = FALSE
)

Arguments

filename

string file name with data to be imported

outputfile

file name and path to store imported data

outGroup

group name to store the dataset

outDataset

dataset name to store the input file in hdf5

sep

(optional), by default = "\t". The field separator string. Values within each row of x are separated by this string.

header

(optional) either a logical value indicating whether the column names of x are to be written along with x, or a character vector of column names to be written. See the section on ‘CSV files’ for the meaning of col.names = NA.

rownames

(optional) either a logical value indicating whether the row names of x are to be written along with x, or a character vector of row names to be written.

overwrite

(optional) either a logical value indicating whether the output file can be overwritten or not.

Value

none value returned, data are stored in a dataset inside an hdf5 data file.