Block matrix multiplication with hdf5 datasets

This function performs a block matrix-matrix multiplication with mattrix stored in HDF5 file

Usage

bdblockSum_hdf5(
  filename,
  group,
  a,
  b,
  groupB = NULL,
  block_size = 2048,
  paral = FALSE,
  threads = NULL,
  outgroup = "OUTPUT",
  outdataset = NULL
)

Arguments

filename

string file name where dataset to normalize is stored

group

string with the group name where matrix is stored inside HDF5 file

a

a double matrix.

b

a double matrix.

groupB,

string, (optional) group name where dataset b is stored

block_size

(optional, defalut = 128) block size to make matrix multiplication, if `block_size = 1` no block size is applied (size 1 = 1 element per block)

paral,

(optional, default = FALSE) paral = true --> TO BE IMPLEMENTED

threads

(optional) only if bparal = true, number of concurrent threads in parallelization if threads is null then threads = maximum number of threads available

outgroup

(optional) string with group name where we want to store the result matrix

outdataset

(optional) string with dataset name where we want to store the results

Value

A list with an HDF5 object with numerical matrix and HDF5 file name with results

"file"string with hdf5 file name with result and original matrixs

"dataset"string complete path inside hdf5 file where results are stored.

Examples

library(BigDataStatMeth)
library(rhdf5)

# Sum matrix twice

# Prepare data and functions
X <- matrix(rnorm(200*10), nrow = 10, ncol = 200)
diag(X) <- 0.5

# Create hdf5 data file with  two X matrices ( named X and Y)
bdCreate_hdf5_matrix_file("test_file3.hdf5", X, "data", "X", force = TRUE)
#> 
#> File and dataset has been created
bdAdd_hdf5_matrix(X, "test_file3.hdf5", "data", "Y", force = TRUE)
#> 
#> Dataset has been added to the file

# Update diagonal
diagonal <- bdblockSum_hdf5( filename = "test_file3.hdf5", 
                   group = "data", a = "X", b = "Y", groupB = "data", 
                   block_size = 128, outgroup = "tmp", outdataset = "Z")