Apply function to different datasets inside a group

Apply function to different datasets inside a group

Usage

bdapply_Function_hdf5(
  filename,
  group,
  datasets,
  outgroup,
  func,
  b_group = NULL,
  b_datasets = NULL,
  force = FALSE,
  transp_dataset = FALSE,
  transp_bdataset = FALSE,
  fullMatrix = FALSE,
  byrows = FALSE,
  threads = 2L
)

Arguments

filename,

Character array, indicating the name of the file to create

group,

Character array, indicating the input group where the data set to be imputed is.

datasets,

Character array, indicating the input datasets to be used

outgroup,

Character, array, indicating group where the data set will be saved after imputation if `outgroup` is NULL, output dataset is stored in the same input group.

func,

Character array, function to be applyed : QR to apply bdQR() function to datasets CrossProd to apply bdCrossprod() function to datasets tCrossProd to apply bdtCrossprod() function to datasets invChol to apply bdInvCholesky() function to datasets blockmult to apply matrix multiplication, in that case, we need the datasets to be used defined in b_datasets variable, datasets and b_datasets must be of the same lenght, in that case, the operation is performed according to index, for example, if we have datasets = "A1", "A2", "A3 and b_datasets = "B1", "B2", "B3, the functions performs : A1 A2 CrossProd_double to performs crossprod using two matrices, see blockmult tCrossProd_double to performs transposed crossprod using two matrices, see blockmult solve to solve matrix equation system, see blockmult for parametrization sdmean to get sd and mean from de datasets by cols or rows

b_group,

optional Character array indicating the input group where data are stored when we need a second dataset to operate, for example in functions like matrix multiplication

b_datasets,

optional Character array indicating the input datasets to be used when we need a second dataset in functions like matrix multiplication

force,

optional Boolean if true, previous results in same location inside hdf5 will be overwritten, by default force = false, data was not overwritten.

transp_dataset

optional parameter. Boolean if true we use the transposed dataframe to perform calculus. By default transp_dataset = false, we use the original dataset stored in hdf5 data file. Currently this option is only valid with "blockmult", "CrossProd_double" and "tCrossProd_double"

transp_bdataset

optional parameter. Boolean if true we use the transposed dataframe to perform calculus.By default transp_bdataset = false, we use the original dataset stored in hdf5 data file. Currently this option is only valid with "blockmult", "CrossProd_double" and "tCrossProd_double"

fullMatrix

boolean, optional parameter used in Inverse Cholesky, by default false. If fullMatrix = true, in the hdf5 file the complete matrix is stored. If false, only the lower triangular matrix is saved

byrows

boolean, optional parameter used in sd and mean calculus, by default false. If byrows = true, the sd and mean is computed by columns. If false, sd and mean is computed by rows.

threads

optional parameter. Integer with numbers of threads to be used

Value

Original hdf5 data file with results after apply function to different datasets